+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* Add optional registration method "phase correlation" to Registration module.
 * Add parallelization into Registration module (lib: multiprocessing).
   (Note that the parellilization works on multiple images and not on the analysis of a single image)
 * Add 'OPTIONS' menu into Registration module